+++ src/IList.h
@@ -1,17 +1,19 @@
 #ifndef _I_LIST_
 #define _I_LIST_
 
+#include <cstddef>
+
 class IList {
 	public:
 		virtual ~IList() {};
 
-		virtual unsigned int getSize() const = 0;
-		virtual int& get(const unsigned int pIndex) const = 0;
+		virtual size_t getSize() const = 0;
+		virtual int& get(const size_t pIndex) const = 0;
 		virtual void add(const int pValue) = 0;
-		virtual int remove(const unsigned int pIndex) = 0;
+		virtual int remove(const size_t pIndex) = 0;
 
-		int& operator[](const unsigned int pIndex) { return get(pIndex); };
-		const int& operator[](const unsigned int pIndex) const { return get(pIndex); };
+		int& operator[](const size_t pIndex) { return get(pIndex); };
+		const int& operator[](const size_t pIndex) const { return get(pIndex); };
 };
 
 #endif
